[eslint-plugin] feat: split no-deprecated-components rule
