+++ packages/eslint-plugin/src/rules/no-deprecated-components/index.ts
@@ -1,0 +1,9 @@
+/*
+ * (c) Copyright 2022 Palantir Technologies Inc. All rights reserved.
+ */
+
+export * from "./no-deprecated-components";
+export * from "./no-deprecated-core-components";
+export * from "./no-deprecated-datetime-components";
+export * from "./no-deprecated-select-components";
+export * from "./no-deprecated-timezone-components";
