+++ packages/eslint-plugin/src/rules/no-deprecated-components/no-deprecated-select-components.ts
@@ -1,0 +1,20 @@
+/* !
+ * (c) Copyright 2022 Palantir Technologies Inc. All rights reserved.
+ */
+
+import { createNoDeprecatedComponentsRule } from "./createNoDeprecatedComponentsRule";
+
+/**
+ * This rule is similar to "@blueprintjs/no-deprecated-components", but it only checks for usage
+ * of deprecated components from @blueprintjs/select. This is useful for incremental migration to
+ * newer Blueprint APIs.
+ */
+export const noDeprecatedSelectComponentsRule = createNoDeprecatedComponentsRule(
+    "no-deprecated-select-components",
+    ["@blueprintjs/select"],
+    {
+        MultiSelect: "MultiSelect2",
+        Select: "Select2",
+        Suggest: "Suggest2",
+    },
+);
